--- src/components/ToastUrlCopyTxt/index.jsx
@@ -1,6 +1,0 @@
-const ToastUrlCopy = () => (
-  <div className='fixed bottom-[100px] md:bottom-[60px] left-1/2 w-[167px] h-[42px] px-5 py-3 bg-black rounded-lg text-sm/[18px] font-medium text-white -translate-x-1/2 shadow-2pt animate-slide-up-fade'>
-    URL이 복사되었습니다
-  </div>
-);
-export default ToastUrlCopy;
